Makefile: Major improvement to the Makefile
As the project grows the task of maintain the Makefile becomes increasingly
cumbersome. This commit tries to improve this situation for the future.

Now the query for .c and .h files become much more easy with automatic search in
folder specified in PROJ_DIRS.
A new variable(MACROS) to declare a macro variable to gcc.
And a automatic detection of the session type, to compile the glfw with wayland.

Signed-off-by: Igor Matheus Andrade Torrente <igormtorrente@gmail.com>