Vulkan: Add the main draw loop
Adds the main draw loop and all the infraestructure to handle window resize.

To handle windows resize all the swapchain and render infrastructure need to
be destroyed and recreated. So the patch refactor the vulkan initialization
and destruction.

A new file was added do handle all this process(vk_resource_manager) and
aditional code was added to the window initialization handle cases were
the api does not report the windows resize.

Signed-off-by: Igor Matheus Andrade Torrente <igormtorrente@gmail.com>