Common: New shared code to handle window inicialization and input
This commit move part of the window inicialization and a callback code from
vk_window to a file called window inside common. That can be used by the
OpenGL backend.

It introduces a new file called types, that contains all the data structure
shared bettwen the Vulkan and OpenGL backend.

And adds new code to handle the input from mouse and keyboard.

The overall objective here is share more code bettween backends and add the
input handle.

Signed-off-by: Igor Matheus Andrade Torrente <igormtorrente@gmail.com>