Vulkan: Common: shaders: Add initial implementation of terrain generation
Common:
Add functions to initialize, generate a terrain given a area and a new struct
to hold information about the terrain.
Remove the Model View Projection(MVP) struct that isn't necessary anymore.

Vulkan:
Implement the functions to manage the cubes positions buffer and change
the record_cmd_buffer, resource_manager and vk_types to deal with the
position buffer.
Add a vec3 descriptor to bind the new buffer with the pipeline.

Shaders:
Change the vertex buffer to accept a position vector from a binded buffer
and generate the Model matrix from this vector.

Signed-off-by: Igor Matheus Andrade Torrente <igormtorrente@gmail.com>