Docker: Add Docker folder
This folder contains three new files:

Dockerfile - A recipe to create a docker image with all the projects dependencies,
This will help people compile and run the project.

dock.sh - A script to help with the docker container.

README.md - A markdown text explaining how to build the container and run the script.

Signed-off-by: Igor Matheus Andrade Torrente <igormtorrente@gmail.com>